chore: various fix && change feat: cmake feat: spilt character fix: unified encoding as UTF-8 fix: "error: out-of-line definition of 'attack' does not match any declaration in 'Character'"
